* commit 'd51d22bcf7c0767812b0b1c86458dbf38b93787b': (137 commits)
  GitIgnored more test-* programs that we currently build
  Gerald's suggestion to switch fields
  Minor typo in rozp
  License update according to Arnaud
  Gerald's suggestions incorporated into rozp
  Problem: src-socket and dst-socket should be strings
  Problem: calc total power can't work with arbitrary ranges
  Copyright year update
  Files generated from rozp.xml, application.xml protocols
  Someone was nitpicking ...
  Based on Alenka's idea
  Transport layer and application protocol draft second version
  ci-rc-bios.sh: sleep a little after systemctl
  ci-rc-bios.sh: fixed search order for compiled programs (use PATH/which as the latest option)
  ci-rc-bios.sh: verify if malamute is already running with the same config - do not restart it in this case; do not stop malamute along with our services in stop() function, do it only for the stop OPERATION
  ci-rc-bios.sh: support non-root execution (invoke sudo and two-step config generation) for systemd; consider malamute service for status and stop activity
  Fixed usage of BUILDSUBDIR in ci-rc-bios.sh; merged Tomas PR#380 and my PR#381 fixes; some other fixes for greater script reliability
  Fixed usage of BUILDSUBDIR in some CI scripts
  Transport and application protocol draft.
  redirect errors to STDERR
  ...

+++ tests/persist/test-topology-location-from.cc
@@ -339,10 +339,10 @@ TEST_CASE("Location topology from #5","[db][topology][location][location_topolog
     r2.clear();
 
     r2 = print_frame_to_edges (asset_msg_grps    (cretTopology),start_id, start_type_id, std::string(start_name), std::string(start_device_type_name));
-    log_info ("r1.size before grp = %d ",(int)r1.size());
+    log_info ("r1.size before grp = %zu ", r1.size());
     r1.insert(r2.begin(), r2.end());
     r2.clear();
-    log_info ("r1.size after grp = %d",(int)r1.size());
+    log_info ("r1.size after grp = %zu", r1.size());
 
     log_info ("start print group elements");
     print_frame (asset_msg_grps    (cretTopology),7000 );
